Test cases & inmem tests.
